+++ src/App.vue
@@ -2,24 +2,19 @@
     <div id="app">
         <!-- <nl-input-test /> -->
         <!-- <nl-textarea-test /> -->
-        <!-- <nl-button-test /> -->
+        <nl-button-test />
         <!-- <nl-progress-test /> -->
-        <nl-avatar-test />
+        <!-- <nl-avatar-test /> -->
         <!-- <timeline-test /> -->
     </div>
 </template>
 
 <script setup>
-import { ref } from "vue";
-
-/**
- * Import components
- */
 // import NlInputTest from "./views/NlInputTest.vue";
 // import NlTextareaTest from "./views/NlTextareaTest.vue";
-// import NlButtonTest from "./views/NlButtonTest.vue";
+import NlButtonTest from "./views/nl-button-test.vue";
 // import NlProgressTest from "./views/NlProgressTest.vue";
-import NlAvatarTest from "./views/nl-avatar-test.vue";
+// import NlAvatarTest from "./views/nl-avatar-test.vue";
 // import TimelineTest from "./views/TimelineTest.vue";
 </script>
